Send "Failed" at the last attempt
Added code to send "Failed" at the last attempt, if the number of attempts to assessment is allowed more than one and if spoor sends "Incomplete" on failing the assessment.
For example: If the number of total attempts is 3 and the spoor is set to send "incomplete" on failing the assessment then this version of the assessment extension will send "incomplete" for the first two attempts and "Failed" on the last attempt. This facilitates the LMS to mark the assignment complete as "Failed".